rearrange model context to facilitate cross-model configuration checking
