+++ filigree-cli/src/model/mod.rs
@@ -7,14 +7,10 @@ use std::borrow::Cow;
 
 use convert_case::{Case, Casing};
 use serde::{Deserialize, Serialize};
-use serde_json::json;
 
-use self::field::{
-    Access, FilterableType, ModelField, ModelFieldReference, ReferentialAction, SqlType,
-};
-use crate::{config::Config, model::field::SortableType};
+use self::field::{Access, ModelField, ModelFieldReference, SqlType};
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Clone, Debug)]
 pub struct Model {
     pub name: String,
     /// The plural of [name], if not generated by adding the letter 's' to the end.
@@ -107,210 +103,6 @@ impl Model {
             .unwrap_or_else(|| Cow::Owned(format!("{}s", self.name)))
     }
 
-    pub fn all_fields(&self) -> impl Iterator<Item = Cow<ModelField>> {
-        self.standard_fields()
-            .map(|field| Cow::Owned(field))
-            .chain(self.fields.iter().map(|field| Cow::Borrowed(field)))
-            .chain(self.reference_fields().map(|field| Cow::Owned(field)))
-    }
-
-    pub fn write_payload_struct_fields(&self) -> impl Iterator<Item = Cow<ModelField>> {
-        self.all_fields()
-            .filter(|f| f.owner_access.can_write() && !f.never_read)
-    }
-
-    pub fn template_context(&self, config: &Config) -> serde_json::Value {
-        let full_default_sort_field = self.default_sort_field.as_deref().unwrap_or("-updated_at");
-        let default_sort_field = if full_default_sort_field.starts_with('-') {
-            &full_default_sort_field[1..]
-        } else {
-            full_default_sort_field
-        };
-
-        {
-            let default_field = self.all_fields().find(|f| f.name == default_sort_field);
-            if let Some(default_field) = default_field {
-                if default_field.sortable == SortableType::None {
-                    panic!(
-                        "Model {}: Default sort field {default_sort_field} is not sortable",
-                        self.name
-                    );
-                }
-            } else {
-                panic!(
-                    "Model {}, Default sort field {} does not exist in model {}",
-                    self.name, default_sort_field, self.name
-                );
-            }
-        }
-
-        let predefined_object_id =
-            &["role", "user", "organization"].contains(&self.module_name().as_str());
-
-        let fields = self
-            .all_fields()
-            .map(|field| field.template_context())
-            .collect::<Vec<_>>();
-
-        json!({
-            "name": self.name,
-            "plural": self.plural(),
-            "table": self.table(),
-            "indexes": self.indexes,
-            "global": self.global,
-            "fields": fields,
-            "owner_permission": format!("{}::owner", self.name),
-            "read_permission": format!("{}::read", self.name),
-            "write_permission": format!("{}::write", self.name),
-            "extra_sql": self.extra_sql,
-            "extra_create_table_sql": self.extra_create_table_sql,
-            "pagination": self.pagination,
-            "full_default_sort_field": full_default_sort_field,
-            "default_sort_field": default_sort_field,
-            "id_type": self.object_id_type(),
-            "id_prefix": self.id_prefix(),
-            "predefined_object_id": predefined_object_id,
-            "url_path": self.plural().to_lowercase(),
-            "has_any_endpoints": self.endpoints.any_enabled(),
-            "endpoints": self.endpoints.per_endpoint(),
-            "auth_scope": self.auth_scope.unwrap_or(config.default_auth_scope),
-        })
-    }
-
-    /// The fields that apply to every object
-    fn standard_fields(&self) -> impl Iterator<Item = ModelField> {
-        let org_field = if self.global {
-            None
-        } else {
-            let org_id_nullable = self.name == "User";
-            let org_id_foreign_key = self.name != "User";
-
-            Some(ModelField {
-                name: "organization_id".to_string(),
-                typ: SqlType::Uuid,
-                rust_type: Some("crate::models::organization::OrganizationId".to_string()),
-                nullable: org_id_nullable,
-                unique: false,
-                indexed: true,
-                sortable: field::SortableType::None,
-                filterable: FilterableType::None,
-                extra_sql_modifiers: String::new(),
-                user_access: Access::Read,
-                owner_access: Access::Read,
-                default_sql: String::new(),
-                default_rust: String::new(),
-                never_read: false,
-                fixed: true,
-                previous_name: None,
-                references: org_id_foreign_key.then(|| {
-                    ModelFieldReference::new(
-                        "organizations",
-                        "id",
-                        Some(ReferentialAction::Cascade),
-                    )
-                }),
-            })
-        };
-
-        [
-            Some(ModelField {
-                name: "id".to_string(),
-                typ: SqlType::Uuid,
-                rust_type: Some(self.object_id_type()),
-                nullable: false,
-                unique: false,
-                indexed: false,
-                filterable: FilterableType::Exact,
-                sortable: field::SortableType::None,
-                extra_sql_modifiers: "primary key".to_string(),
-                user_access: Access::Read,
-                owner_access: Access::Read,
-                references: None,
-                default_sql: String::new(),
-                default_rust: String::new(),
-                never_read: false,
-                fixed: true,
-                previous_name: None,
-            }),
-            org_field,
-            Some(ModelField {
-                name: "updated_at".to_string(),
-                typ: SqlType::Timestamp,
-                rust_type: None,
-                nullable: false,
-                unique: false,
-                indexed: false,
-                filterable: FilterableType::Range,
-                sortable: field::SortableType::DefaultDescending,
-                extra_sql_modifiers: String::new(),
-                user_access: Access::Read,
-                owner_access: Access::Read,
-                references: None,
-                default_sql: "now()".to_string(),
-                default_rust: String::new(),
-                never_read: false,
-                fixed: true,
-                previous_name: None,
-            }),
-            Some(ModelField {
-                name: "created_at".to_string(),
-                typ: SqlType::Timestamp,
-                rust_type: None,
-                nullable: false,
-                unique: false,
-                indexed: false,
-                filterable: FilterableType::Range,
-                sortable: field::SortableType::DefaultDescending,
-                extra_sql_modifiers: String::new(),
-                user_access: Access::Read,
-                owner_access: Access::Read,
-                references: None,
-                default_sql: "now()".to_string(),
-                default_rust: String::new(),
-                never_read: false,
-                fixed: true,
-                previous_name: None,
-            }),
-        ]
-        .into_iter()
-        .flatten()
-    }
-
-    pub fn reference_fields(&self) -> impl Iterator<Item = ModelField> {
-        let belongs_to = self.belongs_to.as_ref().map(|belongs_to| {
-            let id_field_name = format!("{}_id", belongs_to.to_case(Case::Snake));
-            ModelField {
-                name: id_field_name,
-                typ: SqlType::Uuid,
-                rust_type: Some(format!("{belongs_to}Id")),
-                nullable: false,
-                unique: false,
-                indexed: true,
-                filterable: FilterableType::None,
-                sortable: field::SortableType::None,
-                extra_sql_modifiers: String::new(),
-                user_access: Access::Write,
-                owner_access: Access::Write,
-                references: Some(ModelFieldReference {
-                    table: "organizations".to_string(),
-                    field: "id".to_string(),
-                    on_delete: Some(ReferentialAction::Cascade),
-                    on_update: None,
-                    deferrable: None,
-                    populate_on_list: false,
-                    populate_on_get: false,
-                }),
-                default_sql: String::new(),
-                default_rust: String::new(),
-                never_read: false,
-                fixed: false,
-                previous_name: None,
-            }
-        });
-
-        [belongs_to].into_iter().flatten()
-    }
-
     /// The base models use this function to merge in fields added from the config.
     /// Fields with the same name as an existing field in the base model will be replaced,
     /// and other fields will be appended to the end.
@@ -354,7 +146,7 @@ impl Model {
     }
 }
 
-#[derive(Deserialize, Serialize, Debug)]
+#[derive(Deserialize, Serialize, Clone, Debug)]
 pub struct Pagination {
     /// Disable pagination completely unless it's explicitly requested.
     pub disable: bool,
@@ -393,7 +185,7 @@ pub enum SqlDialect {
     SQLite,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Clone, Debug)]
 #[serde(untagged)]
 pub enum Endpoints {
     All(bool),
@@ -491,7 +283,7 @@ pub enum ReferenceFetchType {
     Data,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct HasModel {
     /// The name of the child model
     pub model: String,
